Improve lower data formats logic: Adding less cast ops and some bug fixes (#1983)
### Ticket
Closes #1982 

### Problem description
Lower data format logic uses casts to ensure inputs to graph are in
desired data format. However, some unnecessary casts were being added.
Also, there were some bugs in logic for adding casts. Some data format
bugs throughout the stack are also tackled in this PR.


### What's changed
This PR adds:

- [Fewer Cast Nodes] Insert a cast node on every constant node whose
data format differs from the df_override.
- [Configure Output Data Formats] Set the output data format of all
nodes that need it to match df_override.
    
    Bug Fixes:
- Improve to_torch_operands: if operands use different floating-point
dtypes, cast all to the highest precision.
- Conv2d.eval: avoid casting the output of the evaluation function to
float32.
    - Eltwise Binary: remove unnecessary operand casting.
    
    Tests:
- Add tests that reproduced the aforementioned bugs in
forge/test/mlir/test_dataformats.py.
- Enable lower data format support in YOLOv8 and YOLOv10, and remove
related xfails.

This simplifies specifying lower data formats in tests.

Few side changes are also added:
- bugfix: while iterating through graph nodes and adding constants in
loop, there was a segmentation fault hit. This was because we used range
based loop `(for (Node* node : graph->nodes()))` where `graph->nodes()`
returned a reference to vector of nodes. Inside loop we added new nodes
to that vector. If that vector should be reallocated due to expanding,
iterators in for loop wouldn't be valid anymore. Solution - iterate
using vector size before adding any new nodes. Nodes are added to the
back of the vector so this doesn't cause any problem.

+++ forge/test/models/pytorch/vision/yolo/utils/yolo_utils.py
@@ -16,7 +16,9 @@ class YoloWrapper(torch.nn.Module):
 
     def forward(self, image: torch.Tensor):
         y, x = self.model(image)
-        return (y, *x)
+        # Post processing inside model casts output to float32, even though raw output is aligned with image.dtype
+        # Therefore we need to cast it back to image.dtype
+        return (y.to(image.dtype), *x)
 
 
 def load_yolo_model_and_image(url):
